refactor(learning): remove dead analyze pipeline, extract draft rendering
LEARN-1: the statistical 'analyze' pipeline had zero production callers — the observe
hook uses only getProjectId/isLearningEnabled, learning-curator never imports learning.js,
and 'learn analyze' is retired (exits 1). Removed 6 dead exported entry points
(analyzeLearning, analyzeProjectLearning, analyzeGlobalLearning, analyzeOutcomeRepair,
analyzeTranscriptHabits, extractTranscriptHabits) + triggerAutomaticLearning + their
transitively-dead helpers (tool/sequence/scoring clusters) and the dead test blocks.
Kept the regression-guard tests that assert the analyzer is gone.

ARCH-3: extracted the draft-rendering cluster (render*Draft + renderDraft) into
scripts/lib/learning-drafts.js (direct import, no barrel).

learning.js 1611 -> 744 lines; learning-drafts.js 220. Full npm test green; lint clean.
