+++ scripts/lib/learning.js
@@ -3,6 +3,7 @@ const os = require('node:os');
 const path = require('node:path');
 const crypto = require('node:crypto');
 const { readJsonFile, writeJsonFile } = require('./utils');
+const { renderDraft } = require('./learning-drafts');
 
 const VALID_SCOPES = new Set(['project', 'global']);
 const VALID_STATUSES = new Set(['pending', 'approved', 'rejected', 'materialized', 'activated']);
@@ -335,218 +336,6 @@ function getActiveArtifactPaths(candidate) {
   const paths = ARTIFACT_PATHS[candidate.artifact_type](candidate.name).active;
   paths.forEach(assertNormalizedRelativePath);
   return paths;
-}
-
-function oneLine(value) {
-  return String(value || '')
-    .replace(/\s+/g, ' ')
-    .trim();
-}
-
-function renderSkillDraft(candidate) {
-  const description = JSON.stringify(`Use when ${oneLine(candidate.trigger)}`);
-  return `---
-name: ${candidate.name}
-description: ${description}
----
-
-# ${candidate.name}
-
-> Draft artifact only. This file is intentionally inactive until explicitly activated.
-
-Generated from learning candidate: ${candidate.id}
-
-## Trigger
-
-${candidate.trigger}
-
-## Summary
-
-${candidate.summary}
-
-## Workflow
-
-1. Confirm the user's request matches the trigger.
-2. Review the evidence below and adapt the learned behavior to the current task.
-3. Apply the workflow only when it fits the active project context.
-4. Verify the result with the strongest relevant project checks before reporting completion.
-
-## Evidence
-
-${candidate.evidence.map((item) => `- ${item.source}: ${item.reason} (${item.session_id})`).join('\n')}
-`;
-}
-
-function renderSkillTestDraft(candidate) {
-  const safeName = candidate.name.replace(/-/g, '_');
-  return `from pathlib import Path
-
-
-def test_${safeName}_draft_is_not_active():
-    draft = Path(__file__).with_suffix(Path(__file__).suffix + ".draft")
-    assert draft.name.endswith(".draft")
-
-
-def test_${safeName}_draft_frontmatter_mentions_candidate():
-    draft = Path(__file__).parents[2] / "skills" / "${candidate.name}" / "SKILL.md.draft"
-    text = draft.read_text()
-    assert "name: ${candidate.name}" in text
-    assert "candidate: ${candidate.id}" in text or "${candidate.id}" in text
-`;
-}
-
-function renderInstinctDraft(candidate) {
-  return `---
-name: ${candidate.name}
-description: ${JSON.stringify(`Instinct learned from observation: ${oneLine(candidate.summary)}`)}
-artifact_type: instinct
-status: draft
----
-
-# ${candidate.name}
-
-> Draft instinct — inactive until explicitly activated.
-
-Generated from learning candidate: ${candidate.id}
-
-## Trigger
-
-${candidate.trigger}
-
-## Behavior
-
-${candidate.summary}
-
-## Evidence
-
-${candidate.evidence.map((item) => `- ${item.source}: ${item.reason} (${item.session_id})`).join('\n')}
-`;
-}
-
-function renderCommandDraft(candidate) {
-  return `---
-name: ${candidate.name}
-description: ${JSON.stringify(`Use when ${oneLine(candidate.trigger)}`)}
-artifact_type: command
-status: draft
----
-
-# /${candidate.name}
-
-> Draft command — inactive until explicitly activated.
-
-Generated from learning candidate: ${candidate.id}
-
-## Trigger
-
-${candidate.trigger}
-
-## Behavior
-
-${candidate.summary}
-
-## Evidence
-
-${candidate.evidence.map((item) => `- ${item.source}: ${item.reason} (${item.session_id})`).join('\n')}
-`;
-}
-
-function renderAgentDraft(candidate) {
-  return `---
-name: ${candidate.name}
-description: ${JSON.stringify(`Use when ${oneLine(candidate.trigger)}`)}
-artifact_type: agent
-status: draft
----
-
-# ${candidate.name} agent
-
-> Draft agent definition — inactive until explicitly activated.
-
-Generated from learning candidate: ${candidate.id}
-
-## Trigger
-
-${candidate.trigger}
-
-## Mission
-
-${candidate.summary}
-
-## Evidence
-
-${candidate.evidence.map((item) => `- ${item.source}: ${item.reason} (${item.session_id})`).join('\n')}
-`;
-}
-
-function renderEvalDraft(candidate) {
-  return `---
-name: ${candidate.name}
-description: ${JSON.stringify(`Eval scaffold for ${oneLine(candidate.summary)}`)}
-artifact_type: eval
-status: draft
----
-
-# ${candidate.name} eval
-
-> Draft eval — inactive until explicitly activated.
-
-Generated from learning candidate: ${candidate.id}
-
-## Hypothesis
-
-${candidate.summary}
-
-## Trigger
-
-${candidate.trigger}
-
-## Evidence
-
-${candidate.evidence.map((item) => `- ${item.source}: ${item.reason} (${item.session_id})`).join('\n')}
-`;
-}
-
-function renderRepoConventionPatchDraft(candidate) {
-  // Draft is a human-readable proposal, not an actual unified diff. Activation
-  // is intentionally refused for this type — the user must apply it manually
-  // after review.
-  return `# Proposed repo convention patch (draft only — apply manually)
-
-Candidate: ${candidate.id}
-Target: ${candidate.target || 'AGENTS.md'}
-
-## Trigger
-
-${candidate.trigger}
-
-## Proposed change
-
-${candidate.summary}
-
-## Evidence
-
-${candidate.evidence.map((item) => `- ${item.source}: ${item.reason} (${item.session_id})`).join('\n')}
-`;
-}
-
-function renderDraft(candidate) {
-  switch (candidate.artifact_type) {
-    case 'skill':
-      return [renderSkillDraft(candidate), renderSkillTestDraft(candidate)];
-    case 'instinct':
-      return [renderInstinctDraft(candidate)];
-    case 'command':
-      return [renderCommandDraft(candidate)];
-    case 'agent':
-      return [renderAgentDraft(candidate)];
-    case 'eval':
-      return [renderEvalDraft(candidate)];
-    case 'repo_convention_patch':
-      return [renderRepoConventionPatchDraft(candidate)];
-    default:
-      throw new Error(`unsupported artifact_type: ${candidate.artifact_type}`);
-  }
 }
 
 function materializeCandidate(
@@ -928,657 +717,6 @@ function acceptCandidate(
   throw new Error('candidate must be pending, approved, or materialized to accept');
 }
 
-function readJsonLines(filePath) {
-  if (!fs.existsSync(filePath)) return [];
-  return fs
-    .readFileSync(filePath, 'utf8')
-    .split('\n')
-    .filter(Boolean)
-    .map((line) => {
-      try {
-        return JSON.parse(line);
-      } catch {
-        return null;
-      }
-    })
-    .filter(Boolean);
-}
-
-function listObservationFiles(observationPath) {
-  const files = [];
-  if (fs.existsSync(observationPath)) files.push(observationPath);
-
-  const archiveDir = path.join(path.dirname(observationPath), 'archive');
-  if (fs.existsSync(archiveDir)) {
-    for (const entry of fs.readdirSync(archiveDir).sort()) {
-      if (entry.endsWith('.jsonl')) files.push(path.join(archiveDir, entry));
-    }
-  }
-  return files;
-}
-
-function readObservationFiles(observationPath) {
-  return listObservationFiles(observationPath).flatMap((filePath) => readJsonLines(filePath));
-}
-
-function getObservationRoot({ homeDir } = {}) {
-  return path.join(homePath(homeDir), '.arcforge', 'observations');
-}
-
-function normalizeToolName(tool) {
-  return oneLine(tool || 'unknown') || 'unknown';
-}
-
-function toolSlugComponent(tool) {
-  return (
-    normalizeToolName(tool)
-      .toLowerCase()
-      .replace(/[^a-z0-9]+/g, '-')
-      .replace(/^-+|-+$/g, '') || 'unknown'
-  );
-}
-
-function toolDisplayName(tool) {
-  const words = normalizeToolName(tool)
-    .replace(/[_-]+/g, ' ')
-    .replace(/[^a-z0-9 ]+/gi, ' ')
-    .replace(/\s+/g, ' ')
-    .trim();
-  if (!words) return 'Unknown';
-  return words
-    .split(' ')
-    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
-    .join(' ');
-}
-
-function sequenceKey(tools) {
-  return JSON.stringify(tools.map(toolSlugComponent));
-}
-
-function sequenceSlug(tools) {
-  return tools.map(toolSlugComponent).join('-');
-}
-
-function sequenceStableSlug(tools) {
-  const key = sequenceKey(tools);
-  return `${sequenceSlug(tools)}-${shortHash(key)}`;
-}
-
-function sequenceDisplay(tools) {
-  return tools.map(toolDisplayName).join(' → ');
-}
-
-function projectIdentity(observation, fallback) {
-  return oneLine(
-    observation.__arcforge_project_id ||
-      observation.project_id ||
-      observation.__arcforge_project ||
-      fallback,
-  );
-}
-
-function observationProjectName(observation, fallback) {
-  return oneLine(
-    observation.__arcforge_project || fallback || observation.project || 'unknown-project',
-  );
-}
-
-function sessionKey(observation, { includeProjectIdentity = false } = {}) {
-  const key = oneLine(observation.session || observation.session_id || 'unknown');
-  if (!includeProjectIdentity) return key;
-  return `${projectIdentity(observation, observationProjectName(observation))}:${key}`;
-}
-
-function startedToolObservations(observations) {
-  return observations
-    .filter((observation) => observation.event === 'tool_start' && observation.tool)
-    .sort((a, b) => String(a.ts || '').localeCompare(String(b.ts || '')));
-}
-
-function groupBySession(observations, options = {}) {
-  const groups = new Map();
-  for (const observation of observations) {
-    const key = sessionKey(observation, options);
-    if (!groups.has(key)) groups.set(key, []);
-    groups.get(key).push(observation);
-  }
-  return groups;
-}
-
-function collapseConsecutiveTools(observations) {
-  const tools = [];
-  for (const observation of observations) {
-    const tool = normalizeToolName(observation.tool);
-    if (toolSlugComponent(tools[tools.length - 1]) !== toolSlugComponent(tool)) tools.push(tool);
-  }
-  return tools;
-}
-
-function shortHash(value) {
-  return crypto.createHash('sha256').update(value).digest('hex').slice(0, 8);
-}
-
-function assignUniqueSlugs(patterns) {
-  const counts = new Map();
-  for (const pattern of patterns) {
-    counts.set(pattern.slug, (counts.get(pattern.slug) || 0) + 1);
-  }
-  return patterns.map((pattern) => ({
-    ...pattern,
-    slug: counts.get(pattern.slug) > 1 ? `${pattern.slug}-${shortHash(pattern.key)}` : pattern.slug,
-  }));
-}
-
-function collectWorkflowPatterns(
-  observations,
-  { minSessions = 2, requireDistinctProjects = false, includeProjectInSession = false } = {},
-) {
-  const patterns = new Map();
-  const sessions = groupBySession(startedToolObservations(observations), {
-    includeProjectIdentity: includeProjectInSession,
-  });
-
-  for (const [sessionId, sessionObservations] of sessions.entries()) {
-    const tools = collapseConsecutiveTools(sessionObservations).slice(0, 4);
-    if (tools.length < 2) continue;
-
-    const key = sequenceKey(tools);
-    if (!patterns.has(key)) {
-      patterns.set(key, {
-        key,
-        slug: sequenceStableSlug(tools),
-        tools,
-        display: sequenceDisplay(tools),
-        sessions: new Map(),
-        projects: new Set(),
-      });
-    }
-
-    const first = sessionObservations[0];
-    const pattern = patterns.get(key);
-    pattern.sessions.set(sessionId, first);
-    pattern.projects.add(projectIdentity(first, observationProjectName(first)));
-  }
-
-  const sorted = [...patterns.values()]
-    .filter((pattern) => pattern.sessions.size >= minSessions)
-    .filter((pattern) => !requireDistinctProjects || pattern.projects.size >= 2)
-    .sort((a, b) => {
-      if (b.projects.size !== a.projects.size) return b.projects.size - a.projects.size;
-      if (b.sessions.size !== a.sessions.size) return b.sessions.size - a.sessions.size;
-      if (b.tools.length !== a.tools.length) return b.tools.length - a.tools.length;
-      return a.key.localeCompare(b.key);
-    });
-
-  return assignUniqueSlugs(sorted);
-}
-
-function latestTimestampScore(pattern, now) {
-  const timestamps = [...pattern.sessions.values()]
-    .map((observation) => Date.parse(observation.ts || ''))
-    .filter((value) => Number.isFinite(value));
-  if (timestamps.length === 0) return 0.5;
-  const nowMs = Number.isFinite(Date.parse(now || '')) ? Date.parse(now) : Date.now();
-  const newestMs = Math.max(...timestamps);
-  const ageDays = Math.max(0, (nowMs - newestMs) / 86_400_000);
-  return Math.max(0, 1 - ageDays / 90);
-}
-
-function patternOutcomeScore(pattern) {
-  const outcomes = [...pattern.sessions.values()]
-    .map((observation) => observation.outcome)
-    .filter(Boolean);
-  if (outcomes.length === 0) return 0.5;
-  const successes = outcomes.filter((outcome) => outcome === 'success').length;
-  const errors = outcomes.filter((outcome) => outcome === 'error').length;
-  return Math.max(
-    0,
-    Math.min(1, 0.5 + successes / outcomes.length / 2 - errors / outcomes.length / 2),
-  );
-}
-
-function userConfirmationScore(pattern) {
-  return [...pattern.sessions.values()].some((observation) => observation.user_confirmed === true)
-    ? 1
-    : 0.5;
-}
-
-function contradictionScore(pattern) {
-  return [...pattern.sessions.values()].some(
-    (observation) => observation.contradiction === true || observation.negative_evidence === true,
-  )
-    ? 0
-    : 1;
-}
-
-function projectSpecificityScore(pattern) {
-  // Higher diversity means lower project-specificity risk.
-  return Math.min(1, pattern.projects.size / 3);
-}
-
-function privacyRiskScore(pattern) {
-  const hasRawPayload = [...pattern.sessions.values()].some(
-    (observation) =>
-      typeof observation.input === 'string' ||
-      typeof observation.output === 'string' ||
-      typeof observation.command === 'string',
-  );
-  return hasRawPayload ? 0.2 : 1;
-}
-
-function computeGlobalScore(pattern, { now }) {
-  const factors = {
-    project_diversity: Math.min(1, pattern.projects.size / 3),
-    session_diversity: Math.min(1, pattern.sessions.size / 6),
-    recency: latestTimestampScore(pattern, now),
-    outcome: patternOutcomeScore(pattern),
-    user_confirmation: userConfirmationScore(pattern),
-    contradiction_absence: contradictionScore(pattern),
-    project_specificity: projectSpecificityScore(pattern),
-    privacy_risk: privacyRiskScore(pattern),
-  };
-  const score =
-    factors.project_diversity * 0.22 +
-    factors.session_diversity * 0.18 +
-    factors.recency * 0.14 +
-    factors.outcome * 0.12 +
-    factors.user_confirmation * 0.1 +
-    factors.contradiction_absence * 0.1 +
-    factors.project_specificity * 0.07 +
-    factors.privacy_risk * 0.07;
-  return {
-    score: Number(Math.min(1, score).toFixed(3)),
-    factors: Object.fromEntries(
-      Object.entries(factors).map(([key, value]) => [key, Number(value.toFixed(3))]),
-    ),
-  };
-}
-
-function buildWorkflowCandidate(pattern, { scope, now }) {
-  const global = scope === 'global';
-  const id = `arc-learned-${scope}-${pattern.slug}-workflow`;
-  const name = global
-    ? `arc-global-${pattern.slug}-workflow`
-    : `arc-learned-${pattern.slug}-workflow`;
-  const evidence = [...pattern.sessions.entries()].map(([sessionId, observation]) => ({
-    session_id: oneLine(observation.session || observation.session_id || sessionId),
-    source: global ? `project:${observationProjectName(observation)}` : 'observation',
-    reason: global
-      ? `Repeated ${pattern.display} workflow appears across projects`
-      : `Repeated ${pattern.display} workflow in project observations`,
-  }));
-  const confidence = Math.min(
-    0.85,
-    0.55 + pattern.sessions.size * 0.05 + pattern.projects.size * 0.05,
-  );
-
-  const scoring = global ? computeGlobalScore(pattern, { now }) : null;
-
-  const candidate = {
-    id,
-    scope,
-    artifact_type: 'skill',
-    name,
-    summary: global
-      ? `Global behavior repeated across ${pattern.projects.size} projects and ${pattern.sessions.size} sessions: ${pattern.display}.`
-      : `Project behavior repeated across ${pattern.sessions.size} sessions: ${pattern.display}.`,
-    trigger: global
-      ? `when work across projects repeatedly follows the ${pattern.display} tool workflow`
-      : `when this project work repeatedly follows the ${pattern.display} tool workflow`,
-    evidence,
-    confidence,
-    status: 'pending',
-    pattern_key: pattern.key,
-    created_at: now,
-    updated_at: now,
-  };
-
-  if (global) {
-    candidate.distinct_project_count = pattern.projects.size;
-    candidate.session_count = pattern.sessions.size;
-    candidate.score = scoring.score;
-    candidate.score_factors = scoring.factors;
-  }
-
-  return candidate;
-}
-
-function appendAnalyzerCandidates(candidates, { scope, projectRoot, homeDir }) {
-  const written = [];
-  for (const candidate of candidates) {
-    const result = appendCandidate(candidate, { scope, projectRoot, homeDir });
-    if (!result.duplicate) written.push(result.candidate);
-  }
-  return written;
-}
-
-function analyzeProjectLearning({ projectRoot = process.cwd(), homeDir, now }) {
-  const projectId = getProjectId(projectRoot);
-  const observations = readObservationFiles(getObservationPath({ projectRoot, homeDir })).filter(
-    (observation) => observation.project_id === projectId,
-  );
-  const patterns = collectWorkflowPatterns(observations, { minSessions: 2 });
-  const candidates = patterns.map((pattern) =>
-    buildWorkflowCandidate(pattern, { scope: 'project', now }),
-  );
-  const written = appendAnalyzerCandidates(candidates, { scope: 'project', projectRoot, homeDir });
-  return { scope: 'project', enabled: true, emitted: written.length, candidates: written };
-}
-
-function readGlobalObservations({ homeDir } = {}) {
-  const root = getObservationRoot({ homeDir });
-  if (!fs.existsSync(root)) return [];
-
-  const observations = [];
-  for (const projectName of fs.readdirSync(root).sort()) {
-    const projectDir = path.join(root, projectName);
-    if (!fs.statSync(projectDir).isDirectory()) continue;
-    const records = readObservationFiles(path.join(projectDir, 'observations.jsonl'));
-    for (const record of records) {
-      observations.push({
-        ...record,
-        project: projectName,
-        __arcforge_project: projectName,
-        __arcforge_project_id: projectName,
-      });
-    }
-  }
-  return observations;
-}
-
-function analyzeGlobalLearning({ projectRoot = process.cwd(), homeDir, now }) {
-  const observations = readGlobalObservations({ homeDir });
-  const patterns = collectWorkflowPatterns(observations, {
-    minSessions: 2,
-    requireDistinctProjects: true,
-    includeProjectInSession: true,
-  });
-  const candidates = patterns.map((pattern) =>
-    buildWorkflowCandidate(pattern, { scope: 'global', now }),
-  );
-  const written = appendAnalyzerCandidates(candidates, { scope: 'global', projectRoot, homeDir });
-  return { scope: 'global', enabled: true, emitted: written.length, candidates: written };
-}
-
-function analyzeLearning({
-  scope = 'project',
-  projectRoot = process.cwd(),
-  homeDir,
-  now = new Date().toISOString(),
-} = {}) {
-  assertScope(scope);
-  if (!isLearningEnabled({ scope, projectRoot, homeDir })) {
-    return { scope, enabled: false, emitted: 0, candidates: [] };
-  }
-  if (scope === 'global') return analyzeGlobalLearning({ projectRoot, homeDir, now });
-  return analyzeProjectLearning({ projectRoot, homeDir, now });
-}
-
-function triggerAutomaticLearning({
-  projectRoot = process.cwd(),
-  homeDir,
-  now = new Date().toISOString(),
-} = {}) {
-  const project = analyzeLearning({ scope: 'project', projectRoot, homeDir, now });
-  const global = isLearningEnabled({ scope: 'global', projectRoot, homeDir })
-    ? analyzeLearning({ scope: 'global', projectRoot, homeDir, now })
-    : { scope: 'global', enabled: false, emitted: 0, candidates: [] };
-  return { project, global };
-}
-
-// ---------------------------------------------------------------------------
-// Outcome-aware learning
-//
-// Look for sessions where an error tool_end was followed by an Edit and then
-// a Bash success — a small, common "fix-and-rerun" signal. Emits an instinct
-// candidate. All evidence stays bounded ("test command passed after edit").
-// ---------------------------------------------------------------------------
-
-function findOutcomeRepairSessions(observations, { includeProjectIdentity = false } = {}) {
-  const ordered = [...observations].sort((a, b) =>
-    String(a.ts || '').localeCompare(String(b.ts || '')),
-  );
-  const sessions = new Map();
-  for (const obs of ordered) {
-    const key = sessionKey(obs, { includeProjectIdentity });
-    if (!sessions.has(key)) sessions.set(key, []);
-    sessions.get(key).push(obs);
-  }
-
-  const matches = [];
-  for (const [sessionId, events] of sessions.entries()) {
-    let phase = 'await_error';
-    let pendingRerunKind = null;
-    for (const event of events) {
-      const tool = normalizeToolName(event.tool);
-      const toolSlug = toolSlugComponent(tool);
-      if (event.event === 'tool_end' && event.outcome === 'error') {
-        if (phase === 'await_error') phase = 'await_edit';
-        pendingRerunKind = null;
-      } else if (event.event === 'tool_start' && (toolSlug === 'edit' || toolSlug === 'write')) {
-        if (phase === 'await_edit') phase = 'await_success';
-        pendingRerunKind = null;
-      } else if (event.event === 'tool_start' && toolSlug === 'bash') {
-        const commandKind = oneLine(event.semantic?.command_kind || 'unknown');
-        pendingRerunKind = commandKind === 'test' || commandKind === 'build' ? commandKind : null;
-      } else if (event.event === 'tool_end' && event.outcome === 'success' && toolSlug === 'bash') {
-        if (phase === 'await_success' && pendingRerunKind) {
-          matches.push({
-            sessionId,
-            project: observationProjectName(event),
-            project_id: projectIdentity(event, observationProjectName(event)),
-            command_kind: pendingRerunKind,
-            ts: oneLine(event.ts || ''),
-          });
-          break;
-        }
-        pendingRerunKind = null;
-      }
-    }
-  }
-  return matches;
-}
-
-function buildOutcomeRepairCandidate(matches, { scope, now }) {
-  if (matches.length === 0) return null;
-  const slug = 'outcome-repair-edit-then-test';
-  const id = `arc-learned-${scope}-${slug}-instinct`;
-  const evidence = matches.slice(0, 8).map((match) => ({
-    session_id: match.sessionId || 'unknown',
-    source: scope === 'global' ? `project:${match.project}` : 'observation',
-    reason: 'test command passed after edit following error',
-  }));
-  const confidence = Math.min(0.8, 0.5 + 0.05 * matches.length);
-  return {
-    id,
-    scope,
-    artifact_type: 'instinct',
-    name: `arc-learned-${slug}`,
-    summary: `Recurring fix loop: an error was followed by an edit and a successful re-run.`,
-    trigger:
-      'when a tool errored and the next response was an edit followed by a successful test/build command',
-    evidence,
-    confidence,
-    status: 'pending',
-    pattern_key: `outcome:${slug}`,
-    created_at: now,
-    updated_at: now,
-  };
-}
-
-function analyzeOutcomeRepair({
-  scope = 'project',
-  projectRoot = process.cwd(),
-  homeDir,
-  now = new Date().toISOString(),
-} = {}) {
-  assertScope(scope);
-  if (!isLearningEnabled({ scope, projectRoot, homeDir })) {
-    return { scope, enabled: false, emitted: 0, candidates: [] };
-  }
-  const observations =
-    scope === 'global'
-      ? readGlobalObservations({ homeDir })
-      : readObservationFiles(getObservationPath({ projectRoot, homeDir })).filter(
-          (observation) => observation.project_id === getProjectId(projectRoot),
-        );
-  const matches = findOutcomeRepairSessions(observations, {
-    includeProjectIdentity: scope === 'global',
-  });
-  const candidate = buildOutcomeRepairCandidate(matches, { scope, now });
-  const written = candidate
-    ? appendAnalyzerCandidates([candidate], { scope, projectRoot, homeDir })
-    : [];
-  return { scope, enabled: true, emitted: written.length, candidates: written };
-}
-
-// ---------------------------------------------------------------------------
-// Transcript habit extraction
-//
-// Parses a Claude Code transcript JSONL file and emits privacy-safe instinct
-// or repo_convention_patch candidates from corrections / strong preferences.
-// Never persists raw user/assistant text.
-// ---------------------------------------------------------------------------
-
-const HABIT_PATTERNS = [
-  {
-    kind: 'instinct',
-    regex: /\bdon'?t\s+([a-zA-Z][a-zA-Z\s-]{2,40})/i,
-    prefix: 'user correction observed',
-  },
-  {
-    kind: 'instinct',
-    regex: /\bnever\s+([a-zA-Z][a-zA-Z\s-]{2,40})/i,
-    prefix: 'user correction observed',
-  },
-  {
-    kind: 'instinct',
-    regex: /\bstop\s+([a-zA-Z][a-zA-Z\s-]{2,40})/i,
-    prefix: 'user correction observed',
-  },
-  {
-    kind: 'instinct',
-    regex: /\balways\s+([a-zA-Z][a-zA-Z\s-]{2,40})/i,
-    prefix: 'user preference observed',
-  },
-  {
-    kind: 'repo_convention_patch',
-    regex: /\bfrom\s+now\s+on[,:]?\s+([a-zA-Z][a-zA-Z\s-]{2,60})/i,
-    prefix: 'repo convention preference observed',
-  },
-  {
-    kind: 'instinct',
-    regex: /(?:不要|別|不要再)\s*([^\n]{2,80})/u,
-    prefix: 'user correction observed',
-  },
-  {
-    kind: 'repo_convention_patch',
-    regex: /(?:以後|下次|預設)\s*([^\n]{2,80})/u,
-    prefix: 'repo convention preference observed',
-  },
-];
-
-function extractUserMessageText(line) {
-  if (!line) return null;
-  let parsed;
-  try {
-    parsed = JSON.parse(line);
-  } catch {
-    return null;
-  }
-  if (!parsed || parsed.type !== 'user' || !parsed.message) return null;
-  const content = parsed.message.content;
-  if (typeof content === 'string') return content;
-  if (Array.isArray(content)) {
-    return content
-      .filter((block) => block && block.type === 'text' && typeof block.text === 'string')
-      .map((block) => block.text)
-      .join('\n');
-  }
-  return null;
-}
-
-function safeHabitSummary(prefix, fingerprint) {
-  // Do not echo transcript content, even bounded. Persist only a category plus a
-  // short source-event fingerprint so reviewers can correlate duplicate habits
-  // without storing or hashing private conversation text.
-  return `${prefix} (source_fingerprint:${fingerprint})`;
-}
-
-function extractTranscriptHabits(transcriptPath) {
-  if (!fs.existsSync(transcriptPath)) return [];
-  const lines = fs.readFileSync(transcriptPath, 'utf8').split('\n').filter(Boolean);
-  const habits = [];
-  lines.forEach((line, lineIndex) => {
-    const text = extractUserMessageText(line);
-    if (!text) return;
-    HABIT_PATTERNS.forEach(({ kind, regex, prefix }, patternIndex) => {
-      const match = text.match(regex);
-      if (!match) return;
-      const sourceKey = `${kind}:${prefix}:line-${lineIndex}:pattern-${patternIndex}`;
-      const fingerprint = shortHash(sourceKey);
-      const summary = safeHabitSummary(prefix, fingerprint);
-      const slug = `habit-${fingerprint}`;
-      habits.push({ kind, summary, slug, prefix });
-    });
-  });
-  return habits;
-}
-
-function buildHabitCandidate(habit, { scope, sessionId, now }) {
-  const slugBase = `habit-${habit.kind === 'repo_convention_patch' ? 'convention' : 'instinct'}-${habit.slug}`;
-  const id = `arc-learned-${scope}-${slugBase}`;
-  return {
-    id,
-    scope,
-    artifact_type: habit.kind,
-    name: `arc-learned-${slugBase}`,
-    summary: habit.summary,
-    trigger: `when ${habit.prefix} appears in conversation`,
-    evidence: [
-      {
-        session_id: sessionId,
-        source: 'transcript',
-        reason: habit.summary,
-      },
-    ],
-    confidence: 0.6,
-    status: 'pending',
-    pattern_key: `transcript:${habit.kind}:${habit.slug}`,
-    created_at: now,
-    updated_at: now,
-  };
-}
-
-function analyzeTranscriptHabits({
-  transcriptPath,
-  sessionId = 'unknown',
-  scope = 'project',
-  projectRoot = process.cwd(),
-  homeDir,
-  now = new Date().toISOString(),
-} = {}) {
-  assertScope(scope);
-  if (!isLearningEnabled({ scope, projectRoot, homeDir })) {
-    return { scope, enabled: false, emitted: 0, candidates: [] };
-  }
-  if (!transcriptPath) {
-    throw new Error('analyzeTranscriptHabits requires a transcriptPath');
-  }
-  const habits = extractTranscriptHabits(transcriptPath);
-  // Deduplicate by slug+kind so a single transcript run does not flood the queue.
-  const seen = new Set();
-  const candidates = [];
-  for (const habit of habits) {
-    const key = `${habit.kind}:${habit.slug}`;
-    if (seen.has(key)) continue;
-    seen.add(key);
-    candidates.push(buildHabitCandidate(habit, { scope, sessionId, now }));
-  }
-  const written = appendAnalyzerCandidates(candidates, { scope, projectRoot, homeDir });
-  return { scope, enabled: true, emitted: written.length, candidates: written };
-}
-
 module.exports = {
   REQUIRED_CANDIDATE_FIELDS,
   VALID_SCOPES,
@@ -1589,10 +727,6 @@ module.exports = {
   activateCandidate,
   appendCandidate,
   assertCanMaterialize,
-  analyzeLearning,
-  analyzeOutcomeRepair,
-  analyzeTranscriptHabits,
-  extractTranscriptHabits,
   getCandidateQueuePath,
   getLearningConfigPath,
   getObservationPath,
@@ -1606,6 +740,5 @@ module.exports = {
   readLearningConfig,
   setLearningEnabled,
   transitionCandidate,
-  triggerAutomaticLearning,
   validateCandidate,
 };
