+++ tests/scripts/learning.test.js
@@ -4,7 +4,6 @@ const fs = require('node:fs');
 const os = require('node:os');
 const path = require('node:path');
 const { execFileSync } = require('node:child_process');
-const crypto = require('node:crypto');
 
 const {
   acceptCandidate,
@@ -20,7 +19,6 @@ const {
   readLearningConfig,
   setLearningEnabled,
   transitionCandidate,
-  triggerAutomaticLearning,
   validateCandidate,
 } = require('../../scripts/lib/learning');
 
@@ -434,383 +432,11 @@ describe('learning subsystem MVP-1', () => {
     expect(shouldObserve({ projectRoot, homeDir })).toBe(true);
   });
 
-  function writeObservations(records) {
-    const { getProjectId } = require('../../scripts/lib/learning');
-    const observationPath = path.join(
-      homeDir,
-      '.arcforge',
-      'observations',
-      path.basename(projectRoot),
-      'observations.jsonl',
-    );
-    fs.mkdirSync(path.dirname(observationPath), { recursive: true });
-    fs.writeFileSync(
-      observationPath,
-      `${records
-        .map((record) => JSON.stringify({ project_id: getProjectId(projectRoot), ...record }))
-        .join('\n')}\n`,
-      'utf8',
-    );
-    return observationPath;
-  }
-
-  function writeGlobalProjectObservations(projectName, projectPath, records) {
-    const { getProjectId } = require('../../scripts/lib/learning');
-    const observationPath = path.join(
-      homeDir,
-      '.arcforge',
-      'observations',
-      projectName,
-      'observations.jsonl',
-    );
-    fs.mkdirSync(path.dirname(observationPath), { recursive: true });
-    fs.writeFileSync(
-      observationPath,
-      `${records
-        .map((record) => JSON.stringify({ project_id: getProjectId(projectPath), ...record }))
-        .join('\n')}\n`,
-      'utf8',
-    );
-    return observationPath;
-  }
-
   it('does not contain the old release-specific analyzer entrypoint', () => {
     const source = fs.readFileSync(path.join(__dirname, '../../scripts/lib/learning.js'), 'utf8');
 
     expect(source).not.toContain('releaseSignalScore');
     expect(source).not.toContain('buildReleaseCandidate');
-  });
-
-  it('analyzes repeated project behavior into a pending learned workflow candidate', () => {
-    const { analyzeLearning } = require('../../scripts/lib/learning');
-    setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-    writeObservations([
-      {
-        ts: '2026-05-01T00:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'session-workflow-a',
-        project: path.basename(projectRoot),
-        input: JSON.stringify({ file_path: 'scripts/lib/example.js' }),
-      },
-      {
-        ts: '2026-05-01T00:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'session-workflow-a',
-        project: path.basename(projectRoot),
-        input: JSON.stringify({ file_path: 'scripts/lib/example.js' }),
-      },
-      {
-        ts: '2026-05-01T00:02:00Z',
-        event: 'tool_start',
-        tool: 'Bash',
-        session: 'session-workflow-a',
-        project: path.basename(projectRoot),
-        input: JSON.stringify({ command: 'npm run test:scripts -- learning.test.js' }),
-      },
-      {
-        ts: '2026-05-01T01:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'session-workflow-b',
-        project: path.basename(projectRoot),
-        input: JSON.stringify({ file_path: 'scripts/lib/other.js' }),
-      },
-      {
-        ts: '2026-05-01T01:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'session-workflow-b',
-        project: path.basename(projectRoot),
-        input: JSON.stringify({ file_path: 'scripts/lib/other.js' }),
-      },
-      {
-        ts: '2026-05-01T01:02:00Z',
-        event: 'tool_start',
-        tool: 'Bash',
-        session: 'session-workflow-b',
-        project: path.basename(projectRoot),
-        input: JSON.stringify({ command: 'npm run test:scripts -- other.test.js' }),
-      },
-    ]);
-
-    const result = analyzeLearning({
-      scope: 'project',
-      projectRoot,
-      homeDir,
-      now: '2026-05-01T02:00:00Z',
-    });
-
-    expect(result.enabled).toBe(true);
-    expect(result.emitted).toBe(1);
-    expect(result.candidates[0]).toMatchObject({
-      scope: 'project',
-      artifact_type: 'skill',
-      status: 'pending',
-      pattern_key: JSON.stringify(['read', 'edit', 'bash']),
-    });
-    expect(result.candidates[0].name).toMatch(/^arc-learned-read-edit-bash-[a-f0-9]{8}-workflow$/);
-    expect(result.candidates[0].id).toMatch(
-      /^arc-learned-project-read-edit-bash-[a-f0-9]{8}-workflow$/,
-    );
-    expect(result.candidates[0].trigger).toContain('Read → Edit → Bash');
-    expect(result.candidates[0].summary).not.toMatch(/release/i);
-    expect(result.candidates[0].evidence.map((item) => item.session_id).sort()).toEqual([
-      'session-workflow-a',
-      'session-workflow-b',
-    ]);
-    expect(result.candidates[0].evidence[0]).not.toHaveProperty('input');
-
-    const queued = loadCandidates({ scope: 'project', projectRoot, homeDir });
-    expect(queued).toHaveLength(1);
-    expect(queued[0].name).toMatch(/^arc-learned-read-edit-bash-[a-f0-9]{8}-workflow$/);
-  });
-
-  it('reads active and archived observations when analyzing project behavior', () => {
-    const { analyzeLearning } = require('../../scripts/lib/learning');
-    setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-    const observationPath = writeObservations([
-      {
-        ts: '2026-05-01T00:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'session-archive-a',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T00:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'session-archive-a',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-    ]);
-    const archiveDir = path.join(path.dirname(observationPath), 'archive');
-    fs.mkdirSync(archiveDir, { recursive: true });
-    fs.writeFileSync(
-      path.join(archiveDir, 'observations-2026-05-01.jsonl'),
-      `${[
-        {
-          project_id: require('../../scripts/lib/learning').getProjectId(projectRoot),
-          ts: '2026-05-01T01:00:00Z',
-          event: 'tool_start',
-          tool: 'Read',
-          session: 'session-archive-b',
-          project: path.basename(projectRoot),
-          input: '{}',
-        },
-        {
-          project_id: require('../../scripts/lib/learning').getProjectId(projectRoot),
-          ts: '2026-05-01T01:01:00Z',
-          event: 'tool_start',
-          tool: 'Edit',
-          session: 'session-archive-b',
-          project: path.basename(projectRoot),
-          input: '{}',
-        },
-      ]
-        .map((record) => JSON.stringify(record))
-        .join('\n')}\n`,
-      'utf8',
-    );
-
-    const result = analyzeLearning({ scope: 'project', projectRoot, homeDir });
-
-    expect(result.emitted).toBe(1);
-    expect(result.candidates[0].name).toMatch(/^arc-learned-read-edit-[a-f0-9]{8}-workflow$/);
-    expect(result.candidates[0].evidence.map((item) => item.session_id).sort()).toEqual([
-      'session-archive-a',
-      'session-archive-b',
-    ]);
-  });
-
-  it('does not emit candidates when learning is disabled or evidence is below threshold', () => {
-    const { analyzeLearning } = require('../../scripts/lib/learning');
-    writeObservations([
-      {
-        ts: '2026-05-01T00:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'session-workflow-a',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T00:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'session-workflow-a',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-    ]);
-
-    expect(analyzeLearning({ scope: 'project', projectRoot, homeDir }).enabled).toBe(false);
-    expect(loadCandidates({ scope: 'project', projectRoot, homeDir })).toHaveLength(0);
-
-    setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-    const result = analyzeLearning({ scope: 'project', projectRoot, homeDir });
-
-    expect(result.enabled).toBe(true);
-    expect(result.emitted).toBe(0);
-    expect(loadCandidates({ scope: 'project', projectRoot, homeDir })).toHaveLength(0);
-  });
-
-  it('skips malformed observation lines instead of failing analyzer runs', () => {
-    const { analyzeLearning } = require('../../scripts/lib/learning');
-    setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-    const observationPath = writeObservations([
-      {
-        ts: '2026-05-01T00:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'session-workflow-a',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T00:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'session-workflow-a',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T01:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'session-workflow-b',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T01:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'session-workflow-b',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-    ]);
-    fs.appendFileSync(observationPath, '{not-json}\n', 'utf8');
-
-    const result = analyzeLearning({ scope: 'project', projectRoot, homeDir });
-
-    expect(result.emitted).toBe(1);
-    expect(loadCandidates({ scope: 'project', projectRoot, homeDir })).toHaveLength(1);
-  });
-
-  it('ignores observations from another project id even when project basenames collide', () => {
-    const { analyzeLearning, getProjectId } = require('../../scripts/lib/learning');
-    setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-    const otherProjectRoot = path.join(testDir, 'other', path.basename(projectRoot));
-    writeObservations([
-      {
-        project_id: getProjectId(otherProjectRoot),
-        ts: '2026-05-01T00:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'session-workflow-a',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        project_id: getProjectId(otherProjectRoot),
-        ts: '2026-05-01T00:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'session-workflow-a',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        project_id: getProjectId(otherProjectRoot),
-        ts: '2026-05-01T01:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'session-workflow-b',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        project_id: getProjectId(otherProjectRoot),
-        ts: '2026-05-01T01:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'session-workflow-b',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-    ]);
-
-    const result = analyzeLearning({ scope: 'project', projectRoot, homeDir });
-
-    expect(result.emitted).toBe(0);
-    expect(loadCandidates({ scope: 'project', projectRoot, homeDir })).toHaveLength(0);
-  });
-
-  it('does not queue duplicate analyzer candidates across multiple days', () => {
-    const { analyzeLearning } = require('../../scripts/lib/learning');
-    setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-    writeObservations([
-      {
-        ts: '2026-05-01T00:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'session-workflow-a',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T00:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'session-workflow-a',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T01:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'session-workflow-b',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T01:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'session-workflow-b',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-    ]);
-
-    expect(
-      analyzeLearning({
-        scope: 'project',
-        projectRoot,
-        homeDir,
-        now: '2026-05-01T02:00:00Z',
-      }).emitted,
-    ).toBe(1);
-    expect(
-      analyzeLearning({
-        scope: 'project',
-        projectRoot,
-        homeDir,
-        now: '2026-05-02T02:00:00Z',
-      }).emitted,
-    ).toBe(0);
-
-    const queued = loadCandidates({ scope: 'project', projectRoot, homeDir });
-    expect(queued).toHaveLength(1);
-    expect(queued[0].id).toMatch(/^arc-learned-project-read-edit-[a-f0-9]{8}-workflow$/);
   });
 
   it('CLI learn analyze is deprecated and exits non-zero with a dashboard pointer', () => {
@@ -833,354 +459,6 @@ describe('learning subsystem MVP-1', () => {
     // After deprecation, the analyzer must not silently enqueue candidates.
     expect(loadCandidates({ scope: 'project', projectRoot, homeDir })).toHaveLength(0);
   });
-
-  it('global analyzer emits cross-project learned behavior candidates', () => {
-    const { analyzeLearning, getProjectId } = require('../../scripts/lib/learning');
-    const alphaRoot = path.join(testDir, 'alpha');
-    const betaRoot = path.join(testDir, 'beta');
-    const recordsByProject = [
-      { root: alphaRoot, project: 'alpha', session: 'alpha-session' },
-      { root: betaRoot, project: 'beta', session: 'beta-session' },
-    ];
-    for (const item of recordsByProject) {
-      const dir = path.join(homeDir, '.arcforge', 'observations', item.project);
-      fs.mkdirSync(dir, { recursive: true });
-      fs.writeFileSync(
-        path.join(dir, 'observations.jsonl'),
-        `${[
-          {
-            project_id: getProjectId(item.root),
-            ts: '2026-05-01T00:00:00Z',
-            event: 'tool_start',
-            tool: 'Read',
-            session: item.session,
-            project: item.project,
-            input: '{}',
-          },
-          {
-            project_id: getProjectId(item.root),
-            ts: '2026-05-01T00:01:00Z',
-            event: 'tool_start',
-            tool: 'Edit',
-            session: item.session,
-            project: item.project,
-            input: '{}',
-          },
-        ]
-          .map((record) => JSON.stringify(record))
-          .join('\n')}\n`,
-        'utf8',
-      );
-    }
-    setLearningEnabled({ scope: 'global', enabled: true, projectRoot, homeDir });
-
-    const result = analyzeLearning({ scope: 'global', projectRoot, homeDir });
-
-    expect(result.enabled).toBe(true);
-    expect(result.emitted).toBe(1);
-    expect(result.candidates[0]).toMatchObject({
-      scope: 'global',
-      status: 'pending',
-      pattern_key: JSON.stringify(['read', 'edit']),
-    });
-    expect(result.candidates[0].name).toMatch(/^arc-global-read-edit-[a-f0-9]{8}-workflow$/);
-    expect(result.candidates[0].summary).toContain('2 projects');
-    expect(result.candidates[0].evidence.map((item) => item.session_id).sort()).toEqual([
-      'alpha-session',
-      'beta-session',
-    ]);
-    expect(loadCandidates({ scope: 'global', projectRoot, homeDir })).toHaveLength(1);
-  });
-
-  it('global analyzer trusts observation directory identity over spoofable record project fields', () => {
-    const { analyzeLearning } = require('../../scripts/lib/learning');
-    const alphaRoot = path.join(testDir, 'alpha');
-    writeGlobalProjectObservations('alpha', alphaRoot, [
-      {
-        ts: '2026-05-01T00:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'alpha-session-a',
-        project: 'alpha',
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T00:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'alpha-session-a',
-        project: 'alpha',
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T01:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'alpha-session-b',
-        project: 'spoofed-beta',
-        project_id: 'spoofed-beta-project-id',
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T01:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'alpha-session-b',
-        project: 'spoofed-beta',
-        project_id: 'spoofed-beta-project-id',
-        input: '{}',
-      },
-    ]);
-    setLearningEnabled({ scope: 'global', enabled: true, projectRoot, homeDir });
-
-    const result = analyzeLearning({ scope: 'global', projectRoot, homeDir });
-
-    expect(result.emitted).toBe(0);
-    expect(loadCandidates({ scope: 'global', projectRoot, homeDir })).toHaveLength(0);
-  });
-
-  it('global analyzer treats identical session ids in different projects as separate evidence', () => {
-    const { analyzeLearning } = require('../../scripts/lib/learning');
-    writeGlobalProjectObservations('alpha', path.join(testDir, 'alpha'), [
-      {
-        ts: '2026-05-01T00:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'same-session',
-        project: 'alpha',
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T00:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'same-session',
-        project: 'alpha',
-        input: '{}',
-      },
-    ]);
-    writeGlobalProjectObservations('beta', path.join(testDir, 'beta'), [
-      {
-        ts: '2026-05-01T00:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'same-session',
-        project: 'beta',
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T00:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'same-session',
-        project: 'beta',
-        input: '{}',
-      },
-    ]);
-    setLearningEnabled({ scope: 'global', enabled: true, projectRoot, homeDir });
-
-    const result = analyzeLearning({ scope: 'global', projectRoot, homeDir });
-
-    expect(result.emitted).toBe(1);
-    expect(result.candidates[0].evidence.map((item) => item.source).sort()).toEqual([
-      'project:alpha',
-      'project:beta',
-    ]);
-  });
-
-  it('normalizes analyzer-generated tool workflow names so candidates can materialize', () => {
-    const { analyzeLearning } = require('../../scripts/lib/learning');
-    setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-    writeObservations([
-      {
-        ts: '2026-05-01T00:00:00Z',
-        event: 'tool_start',
-        tool: 'mcp__server__tool',
-        session: 'session-a',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T00:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'session-a',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T01:00:00Z',
-        event: 'tool_start',
-        tool: 'mcp__server__tool',
-        session: 'session-b',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T01:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'session-b',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-    ]);
-
-    const result = analyzeLearning({ scope: 'project', projectRoot, homeDir });
-    const candidateId = result.candidates[0].id;
-    transitionCandidate(candidateId, 'approved', { scope: 'project', projectRoot, homeDir });
-    const materialized = materializeCandidate(candidateId, {
-      scope: 'project',
-      projectRoot,
-      homeDir,
-    });
-
-    expect(result.candidates[0].name).toMatch(
-      /^arc-learned-mcp-server-tool-edit-[a-f0-9]{8}-workflow$/,
-    );
-    expect(materialized.draft_paths[0]).toBe(`skills/${result.candidates[0].name}/SKILL.md.draft`);
-  });
-
-  it('keeps distinct tool sequences separate even when their human slugs would collide', () => {
-    const { analyzeLearning } = require('../../scripts/lib/learning');
-    setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-    const records = [];
-    for (const [session, tools] of [
-      ['ab-c-a', ['a-b', 'c']],
-      ['ab-c-b', ['a-b', 'c']],
-      ['a-bc-a', ['a', 'b-c']],
-      ['a-bc-b', ['a', 'b-c']],
-    ]) {
-      tools.forEach((tool) => {
-        records.push({
-          ts: `2026-05-01T00:0${records.length}:00Z`,
-          event: 'tool_start',
-          tool,
-          session,
-          project: path.basename(projectRoot),
-          input: '{}',
-        });
-      });
-    }
-    writeObservations(records);
-
-    const result = analyzeLearning({ scope: 'project', projectRoot, homeDir });
-    const ids = result.candidates.map((candidate) => candidate.id);
-
-    expect(result.emitted).toBe(2);
-    expect(new Set(ids).size).toBe(2);
-    expect(ids.every((id) => id.startsWith('arc-learned-project-a-b-c'))).toBe(true);
-  });
-
-  it('does not duplicate an existing workflow when a slug-colliding pattern appears in a later analysis run', () => {
-    const { analyzeLearning } = require('../../scripts/lib/learning');
-    setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-    const recordsFor = (sessions) => {
-      const records = [];
-      for (const [session, tools] of sessions) {
-        tools.forEach((tool) => {
-          records.push({
-            ts: `2026-05-01T00:${String(records.length).padStart(2, '0')}:00Z`,
-            event: 'tool_start',
-            tool,
-            session,
-            project: path.basename(projectRoot),
-            input: '{}',
-          });
-        });
-      }
-      return records;
-    };
-
-    writeObservations(
-      recordsFor([
-        ['ab-c-a', ['a-b', 'c']],
-        ['ab-c-b', ['a-b', 'c']],
-      ]),
-    );
-    const firstRun = analyzeLearning({ scope: 'project', projectRoot, homeDir });
-    expect(firstRun.emitted).toBe(1);
-    const firstCandidate = loadCandidates({ scope: 'project', projectRoot, homeDir })[0];
-    expect(firstCandidate.id).toMatch(/^arc-learned-project-a-b-c-[a-f0-9]{8}-workflow$/);
-    expect(firstCandidate.pattern_key).toBe(JSON.stringify(['a-b', 'c']));
-
-    writeObservations(
-      recordsFor([
-        ['ab-c-a', ['a-b', 'c']],
-        ['ab-c-b', ['a-b', 'c']],
-        ['a-bc-a', ['a', 'b-c']],
-        ['a-bc-b', ['a', 'b-c']],
-      ]),
-    );
-    const secondRun = analyzeLearning({ scope: 'project', projectRoot, homeDir });
-    const queued = loadCandidates({ scope: 'project', projectRoot, homeDir });
-
-    expect(secondRun.emitted).toBe(1);
-    expect(queued).toHaveLength(2);
-    expect(queued.map((candidate) => candidate.id)).toContain(firstCandidate.id);
-    expect(new Set(queued.map((candidate) => candidate.pattern_key)).size).toBe(2);
-  });
-
-  it('automatically analyzes enabled project and global learning after new observations without activating artifacts', () => {
-    setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-    setLearningEnabled({ scope: 'global', enabled: true, projectRoot, homeDir });
-    writeObservations([
-      {
-        ts: '2026-05-01T00:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'session-workflow-a',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T00:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'session-workflow-a',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T01:00:00Z',
-        event: 'tool_start',
-        tool: 'Read',
-        session: 'session-workflow-b',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-      {
-        ts: '2026-05-01T01:01:00Z',
-        event: 'tool_start',
-        tool: 'Edit',
-        session: 'session-workflow-b',
-        project: path.basename(projectRoot),
-        input: '{}',
-      },
-    ]);
-
-    const result = triggerAutomaticLearning({
-      projectRoot,
-      homeDir,
-      now: '2026-05-01T02:00:00Z',
-    });
-
-    expect(result.project.enabled).toBe(true);
-    expect(result.project.emitted).toBe(1);
-    expect(result.global.enabled).toBe(true);
-    expect(result.global.emitted).toBe(0);
-    const queued = loadCandidates({ scope: 'project', projectRoot, homeDir });
-    expect(queued).toHaveLength(1);
-    expect(queued[0]).toMatchObject({
-      status: 'pending',
-    });
-    expect(queued[0].name).toMatch(/^arc-learned-read-edit-[a-f0-9]{8}-workflow$/);
-    expect(fs.existsSync(path.join(projectRoot, 'skills', queued[0].name, 'SKILL.md'))).toBe(false);
-    expect(fs.existsSync(path.join(projectRoot, 'skills', queued[0].name, 'SKILL.md.draft'))).toBe(
-      false,
-    );
-  });
-
   it('CLI learn review/approve/reject manages candidate lifecycle without deleting evidence', () => {
     appendCandidate(candidate(), { scope: 'project', projectRoot, homeDir });
     const cli = path.join(__dirname, '../../scripts/cli.js');
@@ -2329,440 +1607,6 @@ describe('learning subsystem MVP-2: multi-artifact-type, outcomes, transcripts',
       expect(() =>
         learning.activateCandidate('evil-1', { scope: 'project', projectRoot, homeDir }),
       ).toThrow(/lowercase kebab-case|relative|parent|normalized/i);
-    });
-  });
-
-  describe('global scoring metadata', () => {
-    it('attaches distinct_project_count, session_count, and score to global candidates', () => {
-      const learning = require('../../scripts/lib/learning');
-      const alphaRoot = path.join(testDir, 'alpha');
-      const betaRoot = path.join(testDir, 'beta');
-      for (const item of [
-        { root: alphaRoot, project: 'alpha', session: 'alpha-session' },
-        { root: betaRoot, project: 'beta', session: 'beta-session' },
-      ]) {
-        const dir = path.join(homeDir, '.arcforge', 'observations', item.project);
-        fs.mkdirSync(dir, { recursive: true });
-        fs.writeFileSync(
-          path.join(dir, 'observations.jsonl'),
-          `${[
-            {
-              project_id: learning.getProjectId(item.root),
-              ts: '2026-05-01T00:00:00Z',
-              event: 'tool_start',
-              tool: 'Read',
-              session: item.session,
-              project: item.project,
-            },
-            {
-              project_id: learning.getProjectId(item.root),
-              ts: '2026-05-01T00:01:00Z',
-              event: 'tool_start',
-              tool: 'Edit',
-              session: item.session,
-              project: item.project,
-            },
-          ]
-            .map((r) => JSON.stringify(r))
-            .join('\n')}\n`,
-          'utf8',
-        );
-      }
-      learning.setLearningEnabled({ scope: 'global', enabled: true, projectRoot, homeDir });
-
-      const result = learning.analyzeLearning({ scope: 'global', projectRoot, homeDir });
-      expect(result.candidates).toHaveLength(1);
-      expect(result.candidates[0].distinct_project_count).toBe(2);
-      expect(result.candidates[0].session_count).toBe(2);
-      expect(result.candidates[0].score).toBeGreaterThan(0);
-      expect(result.candidates[0].score_factors).toMatchObject({
-        project_diversity: expect.any(Number),
-        session_diversity: expect.any(Number),
-        recency: expect.any(Number),
-        outcome: expect.any(Number),
-        user_confirmation: expect.any(Number),
-        contradiction_absence: expect.any(Number),
-        project_specificity: expect.any(Number),
-        privacy_risk: expect.any(Number),
-      });
-      expect(result.candidates[0].status).toBe('pending');
-    });
-  });
-
-  describe('outcome-aware learning', () => {
-    it('emits an instinct candidate from an error → edit → success workflow', () => {
-      const learning = require('../../scripts/lib/learning');
-      learning.setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-      const projectId = learning.getProjectId(projectRoot);
-      const observationPath = learning.getObservationPath({ projectRoot, homeDir });
-      fs.mkdirSync(path.dirname(observationPath), { recursive: true });
-      const events = [
-        // session 1 — error then edit then bash success
-        {
-          ts: '2026-05-01T00:00:00Z',
-          event: 'tool_end',
-          tool: 'Bash',
-          outcome: 'error',
-          session: 's1',
-          project_id: projectId,
-        },
-        {
-          ts: '2026-05-01T00:00:30Z',
-          event: 'tool_start',
-          tool: 'Edit',
-          session: 's1',
-          project_id: projectId,
-        },
-        {
-          ts: '2026-05-01T00:00:45Z',
-          event: 'tool_start',
-          tool: 'Bash',
-          session: 's1',
-          project_id: projectId,
-          semantic: { command_kind: 'test', payload_saved: false },
-        },
-        {
-          ts: '2026-05-01T00:01:00Z',
-          event: 'tool_end',
-          tool: 'Bash',
-          outcome: 'success',
-          session: 's1',
-          project_id: projectId,
-        },
-        // session 2 — same shape
-        {
-          ts: '2026-05-01T01:00:00Z',
-          event: 'tool_end',
-          tool: 'Bash',
-          outcome: 'error',
-          session: 's2',
-          project_id: projectId,
-        },
-        {
-          ts: '2026-05-01T01:00:30Z',
-          event: 'tool_start',
-          tool: 'Edit',
-          session: 's2',
-          project_id: projectId,
-        },
-        {
-          ts: '2026-05-01T01:00:45Z',
-          event: 'tool_start',
-          tool: 'Bash',
-          session: 's2',
-          project_id: projectId,
-          semantic: { command_kind: 'build', payload_saved: false },
-        },
-        {
-          ts: '2026-05-01T01:01:00Z',
-          event: 'tool_end',
-          tool: 'Bash',
-          outcome: 'success',
-          session: 's2',
-          project_id: projectId,
-        },
-      ];
-      fs.writeFileSync(
-        observationPath,
-        `${events.map((e) => JSON.stringify(e)).join('\n')}\n`,
-        'utf8',
-      );
-
-      const result = learning.analyzeOutcomeRepair({
-        scope: 'project',
-        projectRoot,
-        homeDir,
-        now: '2026-05-01T02:00:00Z',
-      });
-
-      expect(result.emitted).toBe(1);
-      const c = result.candidates[0];
-      expect(c.artifact_type).toBe('instinct');
-      expect(c.status).toBe('pending');
-      expect(c.evidence.length).toBeGreaterThanOrEqual(2);
-      // Privacy: no raw payload words allowed in evidence reasons.
-      const evidenceText = JSON.stringify(c.evidence);
-      expect(evidenceText).not.toContain('npm');
-      expect(evidenceText).toMatch(/test command passed after edit/);
-    });
-
-    it('does not emit when there is no error→edit→success sequence', () => {
-      const learning = require('../../scripts/lib/learning');
-      learning.setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-      const projectId = learning.getProjectId(projectRoot);
-      const observationPath = learning.getObservationPath({ projectRoot, homeDir });
-      fs.mkdirSync(path.dirname(observationPath), { recursive: true });
-      fs.writeFileSync(
-        observationPath,
-        `${JSON.stringify({
-          ts: '2026-05-01T00:00:00Z',
-          event: 'tool_start',
-          tool: 'Read',
-          session: 's1',
-          project_id: projectId,
-        })}\n`,
-        'utf8',
-      );
-      const result = learning.analyzeOutcomeRepair({ scope: 'project', projectRoot, homeDir });
-      expect(result.emitted).toBe(0);
-    });
-    it('does not emit when a non-test bash command succeeds after the edit', () => {
-      const learning = require('../../scripts/lib/learning');
-      learning.setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-      const projectId = learning.getProjectId(projectRoot);
-      const observationPath = learning.getObservationPath({ projectRoot, homeDir });
-      fs.mkdirSync(path.dirname(observationPath), { recursive: true });
-      const events = [
-        {
-          ts: '2026-05-01T00:00:00Z',
-          event: 'tool_end',
-          tool: 'Bash',
-          outcome: 'error',
-          session: 's1',
-          project_id: projectId,
-        },
-        {
-          ts: '2026-05-01T00:00:30Z',
-          event: 'tool_start',
-          tool: 'Edit',
-          session: 's1',
-          project_id: projectId,
-        },
-        {
-          ts: '2026-05-01T00:00:45Z',
-          event: 'tool_start',
-          tool: 'Bash',
-          session: 's1',
-          project_id: projectId,
-          semantic: { command_kind: 'git', payload_saved: false },
-        },
-        {
-          ts: '2026-05-01T00:01:00Z',
-          event: 'tool_end',
-          tool: 'Bash',
-          outcome: 'success',
-          session: 's1',
-          project_id: projectId,
-        },
-      ];
-      fs.writeFileSync(
-        observationPath,
-        `${events.map((e) => JSON.stringify(e)).join('\n')}\n`,
-        'utf8',
-      );
-      const result = learning.analyzeOutcomeRepair({ scope: 'project', projectRoot, homeDir });
-      expect(result.emitted).toBe(0);
-    });
-
-    it('does not merge same session ids across projects in global outcome learning', () => {
-      const learning = require('../../scripts/lib/learning');
-      learning.setLearningEnabled({ scope: 'global', enabled: true, projectRoot, homeDir });
-      const root = path.join(homeDir, '.arcforge', 'observations');
-      const alphaPath = path.join(root, 'alpha', 'observations.jsonl');
-      const betaPath = path.join(root, 'beta', 'observations.jsonl');
-      fs.mkdirSync(path.dirname(alphaPath), { recursive: true });
-      fs.mkdirSync(path.dirname(betaPath), { recursive: true });
-      fs.writeFileSync(
-        alphaPath,
-        `${[
-          {
-            ts: '2026-05-01T00:00:00Z',
-            event: 'tool_end',
-            tool: 'Bash',
-            outcome: 'error',
-            session: 'same',
-          },
-          { ts: '2026-05-01T00:00:30Z', event: 'tool_start', tool: 'Edit', session: 'same' },
-        ]
-          .map((e) => JSON.stringify(e))
-          .join('\n')}\n`,
-        'utf8',
-      );
-      fs.writeFileSync(
-        betaPath,
-        `${[
-          {
-            ts: '2026-05-01T00:00:45Z',
-            event: 'tool_start',
-            tool: 'Bash',
-            session: 'same',
-            semantic: { command_kind: 'test', payload_saved: false },
-          },
-          {
-            ts: '2026-05-01T00:01:00Z',
-            event: 'tool_end',
-            tool: 'Bash',
-            outcome: 'success',
-            session: 'same',
-          },
-        ]
-          .map((e) => JSON.stringify(e))
-          .join('\n')}\n`,
-        'utf8',
-      );
-
-      const result = learning.analyzeOutcomeRepair({ scope: 'global', projectRoot, homeDir });
-      expect(result.emitted).toBe(0);
-    });
-  });
-
-  describe('transcript habit extraction', () => {
-    function writeTranscript(lines) {
-      const transcriptPath = path.join(testDir, 'transcript.jsonl');
-      fs.writeFileSync(transcriptPath, lines.map((l) => JSON.stringify(l)).join('\n'), 'utf8');
-      return transcriptPath;
-    }
-
-    it('extracts privacy-safe instinct candidates from corrections in user messages', () => {
-      const learning = require('../../scripts/lib/learning');
-      learning.setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-      const transcriptPath = writeTranscript([
-        { type: 'user', message: { content: "don't push to main without review" } },
-        { type: 'assistant', message: { content: [{ type: 'text', text: 'understood' }] } },
-        { type: 'user', message: { content: 'always run tests before commit' } },
-      ]);
-
-      const result = learning.analyzeTranscriptHabits({
-        transcriptPath,
-        sessionId: 'session-x',
-        scope: 'project',
-        projectRoot,
-        homeDir,
-        now: '2026-05-01T00:00:00Z',
-      });
-
-      expect(result.emitted).toBe(2);
-      const types = result.candidates.map((c) => c.artifact_type);
-      expect(types).toEqual(['instinct', 'instinct']);
-      const reasons = result.candidates.flatMap((c) => c.evidence.map((item) => item.reason));
-      // Evidence reasons stay bounded — no full text echoes.
-      for (const r of reasons) expect(r.length).toBeLessThan(220);
-      for (const c of result.candidates) {
-        expect(c.evidence[0].source).toBe('transcript');
-        expect(c.status).toBe('pending');
-      }
-    });
-
-    it('emits a repo_convention_patch candidate for "from now on" preferences', () => {
-      const learning = require('../../scripts/lib/learning');
-      learning.setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-      const transcriptPath = writeTranscript([
-        { type: 'user', message: { content: 'from now on, prefer kebab-case for filenames' } },
-      ]);
-
-      const result = learning.analyzeTranscriptHabits({
-        transcriptPath,
-        sessionId: 'session-x',
-        scope: 'project',
-        projectRoot,
-        homeDir,
-      });
-
-      expect(result.emitted).toBe(1);
-      expect(result.candidates[0].artifact_type).toBe('repo_convention_patch');
-      expect(result.candidates[0].status).toBe('pending');
-    });
-
-    it('extracts Traditional Chinese correction and preference habits', () => {
-      const learning = require('../../scripts/lib/learning');
-      learning.setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-      const transcriptPath = writeTranscript([
-        { type: 'user', message: { content: '不要自動重啟 Hermes gateway' } },
-        { type: 'user', message: { content: '以後 repo convention 要先提出 AGENTS.md patch' } },
-      ]);
-
-      const result = learning.analyzeTranscriptHabits({
-        transcriptPath,
-        sessionId: 'session-zh',
-        scope: 'project',
-        projectRoot,
-        homeDir,
-      });
-
-      expect(result.emitted).toBe(2);
-      expect(result.candidates.map((c) => c.artifact_type)).toEqual([
-        'instinct',
-        'repo_convention_patch',
-      ]);
-      const serialized = JSON.stringify(result);
-      expect(serialized).not.toContain('Hermes gateway');
-      expect(serialized).not.toContain('AGENTS.md patch');
-    });
-
-    it('does not persist raw transcript text in candidate evidence', () => {
-      const learning = require('../../scripts/lib/learning');
-      learning.setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-      const secret = "don't leak super-secret-token-abcdef123";
-      const transcriptPath = writeTranscript([{ type: 'user', message: { content: secret } }]);
-      const result = learning.analyzeTranscriptHabits({
-        transcriptPath,
-        sessionId: 's1',
-        scope: 'project',
-        projectRoot,
-        homeDir,
-      });
-      const serialized = JSON.stringify(result);
-      expect(serialized).not.toContain('super-secret-token-abcdef123');
-      const contentHash = crypto
-        .createHash('sha256')
-        .update('leak super-secret-token-abcdef123')
-        .digest('hex')
-        .slice(0, 8);
-      expect(serialized).not.toContain(contentHash);
-      expect(result.candidates[0].id).toMatch(
-        /arc-learned-project-habit-instinct-habit-[a-f0-9]{8}$/,
-      );
-      expect(result.candidates[0].pattern_key).toMatch(/^transcript:instinct:habit-[a-f0-9]{8}$/);
-    });
-
-    it('does not derive transcript candidate identifiers from captured text', () => {
-      const learning = require('../../scripts/lib/learning');
-      learning.setLearningEnabled({ scope: 'project', enabled: true, projectRoot, homeDir });
-      const firstTranscript = writeTranscript([
-        { type: 'user', message: { content: "don't expose alpha-secret-value" } },
-      ]);
-      const first = learning.analyzeTranscriptHabits({
-        transcriptPath: firstTranscript,
-        sessionId: 's1',
-        scope: 'project',
-        projectRoot,
-        homeDir,
-      });
-
-      fs.rmSync(path.join(projectRoot, '.arcforge', 'learning', 'candidates'), {
-        recursive: true,
-        force: true,
-      });
-      const secondTranscript = writeTranscript([
-        { type: 'user', message: { content: "don't expose beta-secret-value" } },
-      ]);
-      const second = learning.analyzeTranscriptHabits({
-        transcriptPath: secondTranscript,
-        sessionId: 's2',
-        scope: 'project',
-        projectRoot,
-        homeDir,
-      });
-
-      expect(first.candidates[0].id).toBe(second.candidates[0].id);
-      expect(first.candidates[0].name).toBe(second.candidates[0].name);
-      expect(first.candidates[0].pattern_key).toBe(second.candidates[0].pattern_key);
-    });
-
-    it('does nothing when learning is disabled', () => {
-      const learning = require('../../scripts/lib/learning');
-      const transcriptPath = writeTranscript([
-        { type: 'user', message: { content: 'never use arrow functions' } },
-      ]);
-      const result = learning.analyzeTranscriptHabits({
-        transcriptPath,
-        sessionId: 's1',
-        scope: 'project',
-        projectRoot,
-        homeDir,
-      });
-      expect(result.enabled).toBe(false);
-      expect(result.emitted).toBe(0);
     });
   });
 });
